Move the oracle data broadcast step to steps package
